Fix level selector’s use of digit keys (fix #24)
The level selector was never in an ‘unset’ state, so it was impossible with
the digit keys to select a level that did not start with ‘1’.

Make the ‘level’ variable more explicit: either `None` or a number. When it
is `None`, treat it as 0 for the purposes of adding a digit, or 1 for
incrementing/decrementing.

When a control is pressed, update `level_change` variable rather than
directly updating `level`.

Finally, only change `level` and apply the max/min limits at the end of the
loop if `level_change` is non-zero.